moved number_theory out of ecc submodule
