+++ btclib/ecc/curve_group.py
@@ -22,7 +22,7 @@ from math import ceil
 from typing import List, Sequence, Tuple
 
 from btclib.alias import INF, INFJ, Integer, JacPoint, Point
-from btclib.ecc.number_theory import legendre_symbol, mod_inv, mod_sqrt
+from btclib.number_theory import legendre_symbol, mod_inv, mod_sqrt
 from btclib.exceptions import BTClibTypeError, BTClibValueError
 from btclib.utils import hex_string, int_from_integer
