+++ tests/ecc/test_number_theory.py
@@ -12,7 +12,7 @@
 
 import pytest
 
-from btclib.ecc.number_theory import mod_inv, mod_sqrt, tonelli
+from btclib.number_theory import mod_inv, mod_sqrt, tonelli
 from btclib.exceptions import BTClibValueError
 
 primes = [
