+++ tests/ecc/test_ssa.py
@@ -22,7 +22,7 @@ from btclib.alias import INF, Point, String
 from btclib.bip32.bip32 import BIP32KeyData
 from btclib.ecc import libsecp256k1, ssa
 from btclib.ecc.curve import CURVES, double_mult, mult
-from btclib.ecc.number_theory import mod_inv
+from btclib.number_theory import mod_inv
 from btclib.ecc.pedersen import second_generator
 from btclib.ecc.sec_point import bytes_from_point
 from btclib.exceptions import BTClibRuntimeError, BTClibTypeError, BTClibValueError
